Move bundle filename, add __all__
